Extract peer access + DraftFeatureMirror to src/qwen35/
Move ~250 lines of CUDA peer-access utilities and DraftFeatureMirror
struct/functions from test_dflash.cpp into reusable library files:
- src/qwen35/peer_access.{h,cpp}: enable_peer_access_pair(),
  cross_device_peer_memcpy_ok(), copy_peer_async(), globals
- src/qwen35/draft_feature_mirror.{h,cpp}: DraftFeatureMirror struct,
  init/free/can_view/sync_range/sync_tail functions

test_dflash.cpp now includes these headers with using-declarations.
