+++ dflash/test/test_dflash.cpp
@@ -183,257 +183,21 @@ using dflash27b::StepGraph;
 using dflash27b::step_graph_free;
 using dflash27b::step_graph_destroy;
 
-struct DraftFeatureMirror {
-    ggml_context * ctx = nullptr;
-    ggml_backend_buffer_t buf = nullptr;
-    ggml_tensor * target_feat = nullptr; // F32 [5*hidden, cap]
-    void * bf16_staging = nullptr;
-    size_t bf16_staging_elems = 0;
-    int device = 0;
-    int target_device = 0;
-    int cap = 0;
-};
-
-static bool enable_peer_access_one_way(int device, int peer) {
-    if (device == peer) return true;
-    int can_access = 0;
-    cudaError_t err = cudaDeviceCanAccessPeer(&can_access, device, peer);
-    if (err != cudaSuccess || !can_access) return false;
-    err = cudaSetDevice(device);
-    if (err != cudaSuccess) return false;
-    err = cudaDeviceEnablePeerAccess(peer, 0);
-    if (err == cudaErrorPeerAccessAlreadyEnabled) {
-        cudaGetLastError();
-        return true;
-    }
-    return err == cudaSuccess;
-}
-
-static bool enable_peer_access_pair(int a, int b) {
-    if (a == b) return true;
-    const bool ab = enable_peer_access_one_way(a, b);
-    const bool ba = enable_peer_access_one_way(b, a);
-    return ab && ba;
-}
-
-// Set from argv: opt into cudaMemcpyPeerAsync for cross-device copies when P2P works.
-static bool g_peer_access_opt_in = false;
-static std::unordered_map<std::uint64_t, bool> g_peer_pair_ok_cache;
-
-static std::uint64_t peer_pair_key(int a, int b) {
-    const int lo = std::min(a, b);
-    const int hi = std::max(a, b);
-    return (std::uint64_t)(unsigned)lo << 32 | (unsigned)hi;
-}
-
-static void log_staged_cross_gpu_once() {
-    static bool logged = false;
-    if (logged) {
-        return;
-    }
-    logged = true;
-    std::fprintf(stderr,
-                 "[dflash] Using safe (slower) cross-GPU copy via host staging "
-                 "(--peer-access not set or P2P unavailable for this device pair).\n");
-}
-
-static bool cross_device_peer_memcpy_ok(int src_device, int dst_device) {
-    if (src_device == dst_device) {
-        return true;
-    }
-    if (!g_peer_access_opt_in) {
-        return false;
-    }
-    const std::uint64_t k = peer_pair_key(src_device, dst_device);
-    const auto it = g_peer_pair_ok_cache.find(k);
-    if (it != g_peer_pair_ok_cache.end()) {
-        return it->second;
-    }
-    const bool ok = enable_peer_access_pair(src_device, dst_device);
-    g_peer_pair_ok_cache[k] = ok;
-    return ok;
-}
-
-static bool copy_peer_async(void * dst, int dst_device,
-                            const void * src, int src_device,
-                            size_t bytes,
-                            cudaStream_t stream = nullptr) {
-    if (bytes == 0) return true;
-    cudaError_t err = cudaSuccess;
-    if (dst_device == src_device) {
-        err = cudaSetDevice(dst_device);
-        if (err != cudaSuccess) return false;
-        err = cudaMemcpyAsync(dst, src, bytes, cudaMemcpyDeviceToDevice, stream);
-        if (err != cudaSuccess) return false;
-        if (stream) {
-            return cudaStreamSynchronize(stream) == cudaSuccess;
-        }
-        return cudaDeviceSynchronize() == cudaSuccess;
-    }
-    if (cross_device_peer_memcpy_ok(src_device, dst_device)) {
-        err = cudaSetDevice(dst_device);
-        if (err != cudaSuccess) return false;
-        err = cudaMemcpyPeerAsync(dst, dst_device, src, src_device, bytes, stream);
-        if (err != cudaSuccess) return false;
-        if (stream) {
-            return cudaStreamSynchronize(stream) == cudaSuccess;
-        }
-        return cudaDeviceSynchronize() == cudaSuccess;
-    }
-    log_staged_cross_gpu_once();
-#if defined(DFLASH27B_BACKEND_HIP) || defined(GGML_USE_HIP)
-    err = cudaSetDevice(dst_device);
-    if (err != cudaSuccess) return false;
-    err = cudaMemcpyPeerAsync(dst, dst_device, src, src_device, bytes, stream);
-    if (err != cudaSuccess) return false;
-    if (stream) {
-        return cudaStreamSynchronize(stream) == cudaSuccess;
-    }
-    return cudaDeviceSynchronize() == cudaSuccess;
-#else
-    return dflash_cuda_copy_between_devices(src_device, src, dst_device, dst, bytes,
-                                            nullptr, stream);
-#endif
-}
-
-static bool ensure_bf16_staging(DraftFeatureMirror & mirror, size_t elems) {
-    if (elems <= mirror.bf16_staging_elems) return true;
-    cudaError_t err = cudaSetDevice(mirror.device);
-    if (err != cudaSuccess) return false;
-    if (mirror.bf16_staging) {
-        cudaFree(mirror.bf16_staging);
-        mirror.bf16_staging = nullptr;
-        mirror.bf16_staging_elems = 0;
-    }
-    err = cudaMalloc(&mirror.bf16_staging, elems * sizeof(uint16_t));
-    if (err != cudaSuccess) return false;
-    mirror.bf16_staging_elems = elems;
-    return true;
-}
-
-static void draft_feature_mirror_free(DraftFeatureMirror & mirror) {
-    if (mirror.bf16_staging) {
-        cudaSetDevice(mirror.device);
-        cudaFree(mirror.bf16_staging);
-        mirror.bf16_staging = nullptr;
-        mirror.bf16_staging_elems = 0;
-    }
-    if (mirror.buf) {
-        ggml_backend_buffer_free(mirror.buf);
-        mirror.buf = nullptr;
-    }
-    if (mirror.ctx) {
-        ggml_free(mirror.ctx);
-        mirror.ctx = nullptr;
-    }
-    mirror.target_feat = nullptr;
-    mirror.device = 0;
-    mirror.target_device = 0;
-    mirror.cap = 0;
-}
-
-static bool draft_feature_mirror_init(DraftFeatureMirror & mirror,
-                                      ggml_backend_t backend,
-                                      int device,
-                                      int target_device,
-                                      int cap) {
-    draft_feature_mirror_free(mirror);
-    if (cap <= 0) return false;
-    mirror.device = device;
-    mirror.target_device = target_device;
-
-    ggml_init_params ip{};
-    ip.mem_size = ggml_tensor_overhead() * 4 + 16 * 1024;
-    ip.mem_buffer = nullptr;
-    ip.no_alloc = true;
-    mirror.ctx = ggml_init(ip);
-    if (!mirror.ctx) return false;
-
-    const int fc_in = DFLASH27B_DRAFT_N_TARGET_LAYERS * DFLASH27B_TARGET_HIDDEN;
-    mirror.target_feat = ggml_new_tensor_2d(mirror.ctx, GGML_TYPE_F32, fc_in, cap);
-    ggml_set_name(mirror.target_feat, "draft_target_feat_mirror");
-    mirror.buf = ggml_backend_alloc_ctx_tensors(mirror.ctx, backend);
-    if (!mirror.buf) {
-        draft_feature_mirror_free(mirror);
-        return false;
-    }
-    const size_t bytes = (size_t)fc_in * (size_t)cap * sizeof(float);
-    cudaSetDevice(device);
-    cudaError_t err = cudaMemset(mirror.target_feat->data, 0, bytes);
-    if (err != cudaSuccess) {
-        draft_feature_mirror_free(mirror);
-        return false;
-    }
-    mirror.cap = cap;
-    return true;
-}
-
-static bool draft_feature_mirror_can_view(const DraftFeatureMirror & mirror,
-                                          int committed,
-                                          int ctx_len,
-                                          int & slot0) {
-    if (!mirror.target_feat || mirror.cap <= 0) return false;
-    if (ctx_len <= 0 || ctx_len > mirror.cap || committed < ctx_len) return false;
-    const int start = committed - ctx_len;
-    slot0 = start % mirror.cap;
-    return slot0 + ctx_len <= mirror.cap;
-}
-
-static bool draft_feature_mirror_sync_range(const TargetCache & cache,
-                                            const DraftFeatureMirror & mirror,
-                                            int start_pos,
-                                            int n_tokens) {
-    if (!cache.target_feat || !mirror.target_feat || mirror.cap <= 0) return false;
-    if (n_tokens <= 0) return true;
-    if (n_tokens > mirror.cap) return false;
-
-    const int fc_in = DFLASH27B_DRAFT_N_TARGET_LAYERS * DFLASH27B_TARGET_HIDDEN;
-    const int src_cap = cache.target_feat_cap;
-    const size_t src_stride = cache.target_feat->nb[1];
-    const size_t dst_stride = mirror.target_feat->nb[1];
-
-    int done = 0;
-    while (done < n_tokens) {
-        const int src_slot = (start_pos + done) % src_cap;
-        const int dst_slot = (start_pos + done) % mirror.cap;
-        const int src_run = src_cap - src_slot;
-        const int dst_run = mirror.cap - dst_slot;
-        const int run = std::min(n_tokens - done, std::min(src_run, dst_run));
-        const size_t elems = (size_t)run * (size_t)fc_in;
-        const void * src =
-            (const char *)cache.target_feat->data + (size_t)src_slot * src_stride;
-        void * dst =
-            (char *)mirror.target_feat->data + (size_t)dst_slot * dst_stride;
-        auto bf16_to_f32 = ggml_get_to_fp32_cuda(GGML_TYPE_BF16);
-        if (mirror.device == mirror.target_device) {
-            cudaSetDevice(mirror.device);
-            bf16_to_f32(src, (float *)dst, (int64_t)elems, nullptr);
-        } else {
-            DraftFeatureMirror & mutable_mirror =
-                const_cast<DraftFeatureMirror &>(mirror);
-            if (!ensure_bf16_staging(mutable_mirror, elems)) return false;
-            if (!copy_peer_async(mirror.bf16_staging, mirror.device,
-                                 src, mirror.target_device,
-                                 elems * sizeof(uint16_t))) {
-                return false;
-            }
-            cudaSetDevice(mirror.device);
-            bf16_to_f32(mirror.bf16_staging, (float *)dst, (int64_t)elems, nullptr);
-        }
-        cudaError_t err = cudaGetLastError();
-        if (err != cudaSuccess) return false;
-        done += run;
-    }
-    return cudaDeviceSynchronize() == cudaSuccess;
-}
-
-static bool draft_feature_mirror_sync_tail(const TargetCache & cache,
-                                           const DraftFeatureMirror & mirror,
-                                           int committed) {
-    if (!mirror.target_feat || committed <= 0) return true;
-    const int n = std::min(committed, mirror.cap);
-    return draft_feature_mirror_sync_range(cache, mirror, committed - n, n);
-}
+// ─── Peer access + DraftFeatureMirror — extracted to src/qwen35/ ──
+#include "peer_access.h"
+#include "draft_feature_mirror.h"
+using dflash27b::g_peer_access_opt_in;
+using dflash27b::g_peer_pair_ok_cache;
+using dflash27b::enable_peer_access_one_way;
+using dflash27b::enable_peer_access_pair;
+using dflash27b::cross_device_peer_memcpy_ok;
+using dflash27b::copy_peer_async;
+using dflash27b::DraftFeatureMirror;
+using dflash27b::draft_feature_mirror_free;
+using dflash27b::draft_feature_mirror_init;
+using dflash27b::draft_feature_mirror_can_view;
+using dflash27b::draft_feature_mirror_sync_range;
+using dflash27b::draft_feature_mirror_sync_tail;
 
 // Build a single-layer forward graph for layer-segmented prefill.
 // Processes n_tokens tokens through one layer, reading from act_in and
